Generated JavaDoc files and used Maven to build
The JavaDocs have been placed in a dedicated folder. There was an
attempt to build a workable JAR file for this project using Maven. The
entire main directory was put into a com.comp2042_cw_hcyss4 package
to adhere to the Maven framework. Although there is now a working JAR
file in the target folder, it was most likely was created via InteliJ \
rather than Maven.